// frontend/src/types/profile.ts
export interface Profile {
  fullName: string;
  email: string;
  phone: string;
  profileImage: string;
  college: string;
  degree: string;
  branch: string;
  graduationYear: number;
  cgpa: number;
  skills: string;
  bio: string;
  githubUrl: string;
  linkedinUrl: string;
  portfolioUrl: string;
  resumeUrl: string;
  role: string;
}